address linter warnings
